Remove seperate aarch64c target architecture.
Initially it was though that creating a seperate target
architecture copied from arm64 was the easiest way to handle
code generated for aarch64c/Morello. However, this creates lots of
duplication obscuring the important changes. This would also need
changes to the build scripts. Fianlly, as the simulator must execute
on the actually target architecture there seems little benefit
of adding a new target architecture.