Temprarily prevent checksumming of code for CHERI architectures.
On CHERI a sentry doesn't give ability to read or write the.
The results in a capability fault whne checksumming.

There is seemingly a difficulty here in meeting the requirement
that embedded builtins are both call from a trampoline in non-JIT
builds and remapped into executable memory for JIT builds. It's
not clear how to make both of those cases work, so for the moment
I am just focussed on teh JIT case.
h