Revert changes that allowed static asserts to pass.
Whilst initially attempting to build v8 a large number
of static asserts where failing as a result of the inconsistency
between the frame constants and the use of 64-bit (X) registers
to hold pointer values. As a temporary measure the frame constants
where changed.

As there isn't a clean commit to revert I've had to manually change
the constants back to their original values.